# apps/server/docs/05-openapi.yaml
openapi: 3.1.0
info:
  title: Shafaq Pharmacy Delivery & Consultation API
  version: 1.2.0
  description: |
    Canonical API contract for Shafaq v1.2.
    Backend is the authoritative source of truth for authorization,
    state transitions, ownership, payment, delivery, and consultation privacy.
servers:
  - url: /api/v1
    description: API v1
...
security:
  - bearerAuth: []
paths:
...
  /pharmacies:
    get:
      tags: [Pharmacies]
      security: []
      summary: List pharmacies available to customers
      description: |
        Customer discovery must return pharmacies that are APPROVED and OPEN.
        Internal/admin endpoints may expose additional statuses.
      parameters:
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/RadiusKm'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Pharmacy list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PharmacyListResponse'
...
  /pharmacies/applications:
    post:
      tags: [Pharmacies]
      summary: Submit a pharmacy registration application
      description: |
        Creates a pharmacy with PENDING_APPROVAL status.
        The applicant becomes an OWNER pharmacy membership when accepted
        according to the identity and membership rules.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PharmacyApplicationInput'
      responses:
        '201':
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /pharmacy/assignments/{assignmentId}/accept:
    parameters:
      - $ref: '#/components/parameters/AssignmentId'
    post:
      tags: [Pharmacy Assignments]
      summary: Accept a pharmacy assignment
      responses:
        '200':
          description: Assignment accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /pharmacy/assignments/{assignmentId}/reject:
    parameters:
      - $ref: '#/components/parameters/AssignmentId'
    post:
      tags: [Pharmacy Assignments]
      summary: Reject a pharmacy assignment
      responses:
        '200':
          description: Assignment rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Secure backend session/access token. Production internal users require MFA.
  parameters:
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Latitude:
      name: latitude
      in: query
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
    Longitude:
      name: longitude
      in: query
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
    RadiusKm:
      name: radiusKm
      in: query
      schema:
        type: number
        format: double
        minimum: 0
...
  responses:
...
    Conflict:
      description: Business conflict or concurrent state change
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    SuccessResponse:
      type: object
      required: [success, requestId, data]
      properties:
        success:
          type: boolean
          const: true
        requestId:
          type: string
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [success, requestId, error]
      properties:
        success:
          type: boolean
          const: false
        requestId:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    PharmacyApplicationInput:
      type: object
      required: [name, address, latitude, longitude]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 200
        phone:
          type: string
          maxLength: 20
        address:
          type: string
          maxLength: 500
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    Pharmacy:
      type: object
      required: [id, name, latitude, longitude, approvalStatus, operationalStatus]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        approvalStatus:
          $ref: '#/components/schemas/PharmacyApprovalStatus'
        operationalStatus:
          $ref: '#/components/schemas/PharmacyOperationalStatus'
    PharmacyListResponse:
      type: object
      required: [success, requestId, data]
      properties:
        success:
          type: boolean
          const: true
        requestId:
          type: string
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Pharmacy'
            nextCursor:
              type: string
              nullable: true
...
    Assignment:
      type: object
      required: [id, orderId, pharmacyId, status]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        pharmacyId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AssignmentStatus'
        createdAt:
          type: string
          format: date-time
    AssignmentResponse:
      type: object
      required: [success, requestId, data]
      properties:
        success:
          type: boolean
          const: true
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/Assignment'
...
    PharmacyApprovalStatus:
      type: string
      enum: [PENDING_APPROVAL, APPROVED, REJECTED, SUSPENDED]
    PharmacyOperationalStatus:
      type: string
      enum: [OPEN, CLOSED]
...
    AssignmentStatus:
      type: string
      enum:
        - OFFERED
        - ACTIVE
        - REJECTED
        - EXPIRED
        - CANCELLED
        - COMPLETED
        - TRANSFERRED
...