# apps/server/generated/prisma/browser.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
 * This file should be your main import to use Prisma-related types and utilities in a browser. 
 * Use it to get access to models, enums, and input types.
 * 
 * This file does not contain a `PrismaClient` class, nor several other helpers that are intended as server-side only.
 * See `client.ts` for the standard, server-side entry point.
 *
 * 🟢 You can import this file directly.
 */

import * as Prisma from './internal/prismaNamespaceBrowser.js'
export { Prisma }
export * as $Enums from './enums.js'
export * from './enums.js';
/**
 * Model User
 * 
 */
export type User = Prisma.UserModel
/**
 * Model OtpChallenge
 * 
 */
export type OtpChallenge = Prisma.OtpChallengeModel
/**
 * Model AuthSession
 * 
 */
export type AuthSession = Prisma.AuthSessionModel
/**
 * Model Address
 * 
 */
export type Address = Prisma.AddressModel
/**
 * Model Pharmacy
 * 
 */
export type Pharmacy = Prisma.PharmacyModel
/**
 * Model PharmacyMember
 * 
 */
export type PharmacyMember = Prisma.PharmacyMemberModel
/**
 * Model Driver
 * 
 */
export type Driver = Prisma.DriverModel
/**
 * Model Order
 * 
 */
export type Order = Prisma.OrderModel
/**
 * Model OrderItem
 * 
 */
export type OrderItem = Prisma.OrderItemModel
/**
 * Model PharmacyAssignment
 * 
 */
export type PharmacyAssignment = Prisma.PharmacyAssignmentModel