# apps/server/generated/prisma/client.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
 * This file should be your main import to use Prisma. Through it you get access to all the models, enums, and input types.
 * If you're looking for something you can import in the client-side of your application, please refer to the `browser.ts` file instead.
 *
 * 🟢 You can import this file directly.
 */

import * as process from 'node:process'
import * as path from 'node:path'
import { fileURLToPath } from 'node:url'
globalThis['__dirname'] = path.dirname(fileURLToPath(import.meta.url))

import * as runtime from "@prisma/client/runtime/client"
import * as $Enums from "./enums.js"
import * as $Class from "./internal/class.js"
import * as Prisma from "./internal/prismaNamespace.js"

export * as $Enums from './enums.js'
export * from "./enums.js"
/**
 * ## Prisma Client
 * 
 * Type-safe database client for TypeScript
 * @example
 * ```
 * const prisma = new PrismaClient({
 *   adapter: new PrismaPg({ connectionString: process.env.DATABASE_URL })
 * })
 * // Fetch zero or more Users
 * const users = await prisma.user.findMany()
 * ```
 * 
 * Read more in our [docs](https://pris.ly/d/client).
 */
export const PrismaClient = $Class.getPrismaClientClass()
export type PrismaClient<LogOpts extends Prisma.LogLevel = never, OmitOpts extends Prisma.PrismaClientOptions["omit"] = Prisma.PrismaClientOptions["omit"], ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = $Class.PrismaClient<LogOpts, OmitOpts, ExtArgs>
export { Prisma }

/**
 * Model User
 * 
 */
export type User = Prisma.UserModel
/**
 * Model OtpChallenge
 * 
 */
export type OtpChallenge = Prisma.OtpChallengeModel
/**
 * Model AuthSession
 * 
 */
export type AuthSession = Prisma.AuthSessionModel
/**
 * Model Address
 * 
 */
export type Address = Prisma.AddressModel
/**
 * Model Pharmacy
 * 
 */
export type Pharmacy = Prisma.PharmacyModel
/**
 * Model PharmacyMember
 * 
 */
export type PharmacyMember = Prisma.PharmacyMemberModel
/**
 * Model Driver
 * 
 */
export type Driver = Prisma.DriverModel
/**
 * Model Order
 * 
 */
export type Order = Prisma.OrderModel
/**
 * Model OrderItem
 * 
 */
export type OrderItem = Prisma.OrderItemModel
/**
 * Model PharmacyAssignment
 * 
 */
export type PharmacyAssignment = Prisma.PharmacyAssignmentModel